setting up SERVER for interannotation, NEXT: adding files, computing errors
