+++ server/annotator.py
@@ -4,9 +4,12 @@
 
 # >>>> Native <<<<
 import os
+import sys
 import json
 import copy
+import time
 from typing import Dict, List, Any, Tuple, Hashable, Iterable, Union
+import functools
 
 # >>>> Flask <<<<
 from flask import Flask, jsonify, request
@@ -21,25 +24,19 @@ from utils import load_json_file, save_json_file
 # CODE
 ################################################################################
 
-class Multiannotator(object):
+class MultiAnnotator(object):
     """
     The class that governs several DialogueAnnotator(s)
     """
+    __GOLD_FILE_NAME = "GOLD_" + str( time.strftime("%Y%m%d-%H%M%S") )
     __DEFAULT_NAME = "FILE_"
 
     def __init__(self, path):
         self.path = path
-        self.allFiles = {}
+        self.allFiles = { MultiAnnotator.__GOLD_FILE_NAME : DialogueAnnotator(self.path, MultiAnnotator.__GOLD_FILE_NAME) }
         self.filesAdded = 0
 
-        self.load_all_jsons(self.path)
-
-    def get(self, fileId):
-        """
-        gets the relevant file
-        """
-
-        return self.allFiles.get(fileId)
+        self.__load_all_jsons(self.path)
 
 
     def add_dialogue_file( self, jsonObjectOrFileName ):
@@ -51,6 +48,7 @@ class Multiannotator(object):
         else:
             fileName = self.__get_new_file_id()
             save_json_file( obj=jsonObjectOrFileName, path= os.path.join( self.path, fileName ) )
+            self.allFiles[MultiAnnotator.__GOLD_FILE_NAME].update_dialogues(jsonObjectOrFileName)
 
         self.filesAdded += 1
         self.allFiles[ fileName ] = DialogueAnnotator( self.path, fileName )
@@ -61,6 +59,32 @@ class Multiannotator(object):
         returns the names of the files
         """
         return { "names" : [ {"name":x} for x,_ in self.allFiles.items() ] }
+
+
+
+    def dialogue_file_function_call(self, methodName, fileId=None, **args):
+        """
+        Makes the relevant call from the single dialogue file
+        """
+        if not fileId:
+            fileId = MultiAnnotator.__GOLD_FILE_NAME
+
+        temp = self.allFiles.get( fileId )
+
+        method = getattr(temp, methodName)
+
+        return method(**args)
+
+
+    def __getattr__(self, attr):
+        """
+        Magic Methods++
+        """
+        
+        temp = functools.partial(self.dialogue_file_function_call, attr)
+
+        return temp
+
 
 
     def __load_all_jsons(self,path):
@@ -88,32 +112,45 @@ class Multiannotator(object):
 
 
 
+
+
+
+
 class DialogueAnnotator(object):
     """
     class that handles everything which relates to managing a single dialogues file
     """
     __DEFAULT_FILENAME="labelled_data.json"
 
-    def __init__( self, filePath, fileName=None ):
+    def __init__( self, filePath, fileName=None, dialogues=None ):
         """
         """
+        self.set_dialogues( dialogues )
         self.set_file( filePath, fileName )
         self.addedDialogues = 0
 
 
+    def set_dialogues( self, dialogues=None ):
+        """
+        """
+        self.__dialogues = dialogues if dialogues else {}
 
     def set_file( self, filePath, fileName=None ):
         """
+        sets the file and tries to load it to use
         """
         self.__filePath = filePath
 
         if fileName:
             self.__fileName = fileName
-            self.__dialogues = load_json_file( os.path.join( self.__filePath, self.__fileName ) )
+            try:
+                self.__dialogues = load_json_file( os.path.join( self.__filePath, self.__fileName ) )
+            except FileNotFoundError:
+                save_json_file( obj=self.__dialogues, path=os.path.join( self.__filePath, self.__fileName ) )
 
         else:
             self.__fileName = DialogueAnnotator.__DEFAULT_FILENAME
-            self.__dialogues = {}
+
 
 
     def get_dialogues(self, id=None):
@@ -121,7 +158,7 @@ class DialogueAnnotator(object):
         Returns all dialogues or specific dialogue (as dict {id: dialogue} )
         """
         if id:
-            return {"dialogue" : self.__dialogues[id]}
+            return { "dialogue" : self.__dialogues.get(id) }
 
         return self.__dialogues
 
@@ -149,6 +186,14 @@ class DialogueAnnotator(object):
 
         return {"status" : "success"}
 
+
+    def update_dialogues(self, newDialogues):
+        """
+        updates all the dialogues with a new dictionary
+        """
+        self.__dialogues.update( newDialogues )
+
+        return True
 
 
     def update_dialogue_name(self, id, newName):
